+++ app/lib/util-service.js
@@ -198,6 +198,58 @@ angular.module('lizard-nxt')
     }
   };
 
+
+  // helper for time stuff
+  this.hour = 60 * 60 * 1000;
+  this.day = 24 * this.hour;
+
+  /**
+   * @function parseDaysHours
+   * @param {string} 0Days3Hours string
+   * @return {int} milliseconds representation
+   */
+  this.parseDaysHours = function (timeString) {
+    if (timeString === undefined) {
+      return 0;
+    }
+
+    var days = parseInt(timeString.split('Days')[0]);
+    var hours = parseInt(timeString.split('Days')[1].split('Hours')[0]);
+
+    var totalMS = 0;
+    if (!isNaN(days)) {
+      totalMS += parseInt(days) * this.day;
+    }
+
+    if (!isNaN(hours)) {
+      totalMS += parseInt(hours) * this.hour;
+    }
+
+    return totalMS;
+  };
+
+  /**
+   * @function getTimeIntervalAsText
+   * @param {int} Start time
+   * @param {int} End time
+   * @return {string} Difference in format of: 7 days 3 hours
+   */
+  this.getTimeIntervalAsText = function (start, end) {
+    var days = '',
+        hours = '';
+
+    // only calculate if the end is larger than start
+    if (end > start) {
+      var interval = end - start;
+      days = Math.floor(interval / this.day);
+      hours = Math.floor((interval % this.day) / this.hour);
+    }
+    return {
+      days: days,
+      hours: hours
+    };
+  };
+
   /**
    * @function metersToDegs
    * @memberOf UtilService